Attempting to address Travis and try build failures.
Introduce CFG_USING_CLANG, which is distinguished from
CFG_ENABLE_CLANG because the former represents "we think we're using
clang, choose appropriate warning-control options" while the latter
represents "we asked configure (or the host required) that we attempt
to use clang, so check that we have an appropriate version of clang."

The main reason I added this is that I wanted to allow the user to
choose clang via setting the `CC` environment variable, but I did not
want that method of selection to get confused with the user passing
the `--enable-clang` option.

Drive-by fix: Fix chk_cc error message to point user at `gcc` not `cc`.

----

Also, for some reason these changes are exposing issues where the
builds of native libraries and the compiler-rt were not passing
through the user-selected values for `$CFLAGS`, which causes problems
e.g. when want to silence clang warnings that cause the compilation to
fail (namely about unused command line options); fixed that.